+++ src/Atata.WebDriverSetup/Strategies/FirefoxDriverSetupStrategy.cs
@@ -46,14 +46,14 @@ public class FirefoxDriverSetupStrategy :
         };
 
     /// <inheritdoc/>
-    public string? GetInstalledBrowserVersion() =>
+    public async Task<string?> GetInstalledBrowserVersionAsync(CancellationToken cancellationToken = default) =>
         OSInfo.IsWindows
             ? AppVersionDetector.GetFromProgramFiles(@"Mozilla Firefox\firefox.exe")
                 ?? RegistryUtils.GetValue(@"HKEY_CURRENT_USER\Software\Mozilla\Mozilla Firefox")
                 ?? AppVersionDetector.GetByApplicationPathInRegistry("firefox.exe")
             : (OSInfo.IsOSX
-                ? AppVersionDetector.GetThroughOSXApplicationCli("Firefox")
-                : AppVersionDetector.GetThroughCli("firefox", "-v"))
+                ? (await AppVersionDetector.GetThroughOSXApplicationCliAsync("Firefox", cancellationToken).ConfigureAwait(false))
+                : (await AppVersionDetector.GetThroughCliAsync("firefox", "-v", cancellationToken).ConfigureAwait(false)))
                 ?.Replace("Mozilla Firefox ", null);
 
     /// <inheritdoc/>