+++ test/Atata.WebDriverSetup.IntegrationTests/DriverSetupConfigurationBuilderTests.cs
@@ -7,11 +7,11 @@ public class DriverSetupConfigurationBuilderTests : IntegrationTestFixture
     [TestCase(BrowserNames.Edge)]
     [TestCase(BrowserNames.Opera)]
     [TestCase(BrowserNames.InternetExplorer, IncludePlatform = Platforms.Windows)]
-    public void SetUp_WithAutoVersion(string browserName)
+    public async Task SetUpAsync_WithAutoVersion(string browserName)
     {
-        var result = DriverSetup.Configure(browserName)
+        var result = await DriverSetup.Configure(browserName)
             .WithAutoVersion()
-            .SetUp();
+            .SetUpAsync();
 
         AssertDriverIsSetUp(result, browserName);
         AssertVersionCache(browserName);
@@ -22,12 +22,12 @@ public class DriverSetupConfigurationBuilderTests : IntegrationTestFixture
     [TestCase(BrowserNames.Edge, IncludePlatform = Platforms.Windows)]
     [TestCase(BrowserNames.Opera, IncludePlatform = Platforms.Windows)]
     [TestCase(BrowserNames.InternetExplorer, IncludePlatform = Platforms.Windows)]
-    public void SetUp_WithAutoVersion_WithX32Architecture(string browserName)
+    public async Task SetUpAsync_WithAutoVersion_WithX32Architecture(string browserName)
     {
-        var result = DriverSetup.Configure(browserName)
+        var result = await DriverSetup.Configure(browserName)
             .WithAutoVersion()
             .WithX32Architecture()
-            .SetUp();
+            .SetUpAsync();
 
         AssertDriverIsSetUp(result, browserName, architecture: Architecture.X32);
         AssertVersionCache(browserName);
@@ -38,11 +38,11 @@ public class DriverSetupConfigurationBuilderTests : IntegrationTestFixture
     [TestCase(BrowserNames.Edge, IncludePlatform = Platforms.Windows)]
     [TestCase(BrowserNames.Opera)]
     [TestCase(BrowserNames.InternetExplorer, IncludePlatform = Platforms.Windows)]
-    public void SetUp_WithLatestVersion(string browserName)
+    public async Task SetUpAsync_WithLatestVersion(string browserName)
     {
-        var result = DriverSetup.Configure(browserName)
+        var result = await DriverSetup.Configure(browserName)
             .WithLatestVersion()
-            .SetUp();
+            .SetUpAsync();
 
         AssertDriverIsSetUp(result, browserName);
         AssertVersionCache(browserName);
@@ -58,11 +58,11 @@ public class DriverSetupConfigurationBuilderTests : IntegrationTestFixture
     [TestCase(BrowserNames.Opera, "86.0.4240.80")]
     [TestCase(BrowserNames.InternetExplorer, "3.141.59", IncludePlatform = Platforms.Windows)]
     [TestCase(BrowserNames.InternetExplorer, "4.11.0", IncludePlatform = Platforms.Windows)]
-    public void SetUp_WithVersion(string browserName, string version)
+    public async Task SetUpAsync_WithVersion(string browserName, string version)
     {
-        var result = DriverSetup.Configure(browserName)
+        var result = await DriverSetup.Configure(browserName)
             .WithVersion(version)
-            .SetUp();
+            .SetUpAsync();
 
         AssertDriverIsSetUp(result, browserName, version);
     }
@@ -74,11 +74,11 @@ public class DriverSetupConfigurationBuilderTests : IntegrationTestFixture
     [TestCase(BrowserNames.Chrome, "117.0.5896", "117.0.5896.0")]
     [TestCase(BrowserNames.Firefox, "101", "0.31.0")]
     [TestCase(BrowserNames.Edge, "135.0.3179.54", "135.0.3179.54")]
-    public void SetUp_ByBrowserVersion(string browserName, string browserVersion, string driverVersion)
+    public async Task SetUpAsync_ByBrowserVersion(string browserName, string browserVersion, string driverVersion)
     {
-        var result = DriverSetup.Configure(browserName)
+        var result = await DriverSetup.Configure(browserName)
             .ByBrowserVersion(browserVersion)
-            .SetUp();
+            .SetUpAsync();
 
         AssertDriverIsSetUp(result, browserName, driverVersion);
         AssertVersionCache(browserName, browserVersion);
@@ -86,24 +86,25 @@ public class DriverSetupConfigurationBuilderTests : IntegrationTestFixture
 
     [TestCase(BrowserNames.Opera, "73.0.3856.329")]
     [TestCase(BrowserNames.InternetExplorer, "11.0.0.4", IncludePlatform = Platforms.Windows)]
-    public void SetUp_ByBrowserVersion_Unsupported(string browserName, string version)
+    public void SetUpAsync_ByBrowserVersion_Unsupported(string browserName, string version)
     {
         var builder = DriverSetup.Configure(browserName)
             .ByBrowserVersion(version);
 
-        Assert.Throws<DriverSetupException>(() =>
-            builder.SetUp());
+        Assert.That(
+            async () => await builder.SetUpAsync(),
+            Throws.TypeOf<DriverSetupException>());
     }
 
     [Test]
-    public void SetUp_Edge_WithAutoVersion_WhichFailsToDownloadAndClosestVersionIsDownloaded() =>
-        TestEdgeSetUpWhichFailsToDownloadAndPreviousVersionIsDownloaded(x => x.WithAutoVersion());
+    public async Task SetUpAsync_Edge_WithAutoVersion_WhichFailsToDownloadAndClosestVersionIsDownloaded() =>
+        await TestEdgeSetUpAsyncWhichFailsToDownloadAndPreviousVersionIsDownloadedAsync(x => x.WithAutoVersion());
 
     [Test]
-    public void SetUp_Edge_WithLatestVersion_WhichFailsToDownloadAndClosestVersionIsDownloaded() =>
-        TestEdgeSetUpWhichFailsToDownloadAndPreviousVersionIsDownloaded(x => x.WithLatestVersion());
+    public async Task SetUpAsync_Edge_WithLatestVersion_WhichFailsToDownloadAndClosestVersionIsDownloaded() =>
+        await TestEdgeSetUpAsyncWhichFailsToDownloadAndPreviousVersionIsDownloadedAsync(x => x.WithLatestVersion());
 
-    private static void TestEdgeSetUpWhichFailsToDownloadAndPreviousVersionIsDownloaded(
+    private static async Task TestEdgeSetUpAsyncWhichFailsToDownloadAndPreviousVersionIsDownloadedAsync(
         Action<DriverSetupConfigurationBuilder> builderConfiguration)
     {
         FakeHttpRequestExecutorProxy fakeHttpRequestExecutorProxy = null!;
@@ -112,7 +113,7 @@ public class DriverSetupConfigurationBuilderTests : IntegrationTestFixture
 
         builderConfiguration.Invoke(builder);
 
-        var result = builder
+        var result = await builder
             .WithCheckCertificateRevocationList(false)
             .WithHttpRequestTryCount(1)
             .WithHttpRequestExecutor(
@@ -121,7 +122,7 @@ public class DriverSetupConfigurationBuilderTests : IntegrationTestFixture
                 {
                     DownloadFileFailuresCount = 1
                 })
-            .SetUp();
+            .SetUpAsync();
 
         AssertDriverIsSetUp(result, BrowserNames.Edge);
         AssertVersionCache(BrowserNames.Edge);