+++ test/Atata.WebDriverSetup.IntegrationTests/EdgeDriverSetupStrategyTests.cs
@@ -75,8 +75,8 @@ public class EdgeDriverSetupStrategyTests
     }
 
     [Test]
-    public void GetInstalledBrowserVersion() =>
-        _sut.GetInstalledBrowserVersion()
+    public async Task GetInstalledBrowserVersionAsync() =>
+        (await _sut.GetInstalledBrowserVersionAsync())
             .Should().MatchRegex(@"^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+$");
 
     [TestCase("100.0.1155.0", ExpectedResult = "100.0.1155.0")]